Fix NPE
(cherry picked from commit c57fa9d6084c2b8c9f66e7c25e781607beeb1f3f)

+++ src/main/java/eu/siacs/conversations/parser/MessageParser.java
@@ -909,8 +909,10 @@ public class MessageParser extends AbstractParser implements OnMessagePacketRece
                                             Cid cid = BobTransfer.cid(uri);
                                             if (cid == null) continue;
                                             DownloadableFile f = mXmppConnectionService.getFileForCid(cid);
-                                            mXmppConnectionService.evictPreview(f);
-                                            f.delete();
+                                            if (f != null) {
+                                                mXmppConnectionService.evictPreview(f);
+                                                f.delete();
+                                            }
                                         }
                                     }
                                 }
